requirements.txt
numpy==1.14
geopy==1.10
paho-mqtt==1.3.1
requests==2.18.4
